BREAKING CHANGE: rename stitches.config to config
