+++ package.json
@@ -24,7 +24,7 @@
   },
   "files": [
     "gatsby-ssr.js",
-    "src/stitches.config.ts",
+    "src/config.ts",
     "index.js"
   ],
   "peerDependencies": {
